slight refactor of expense table for splitting row usecases
