+++ src/app/components/ExpenseTable.tsx
@@ -1,11 +1,10 @@
-import { convertToFloat, sortExpenses } from "@/app/lib/helpers";
-import { expense } from "@/app/lib/types";
+import { sortExpenses } from "@/app/lib/helpers";
 import { Expense } from "@prisma/client";
 import { useMemo, useState } from "react";
-import RecurringIcon from "../AddExpense/RecurringSVG";
+import ExpenseRow from "./Dashboard/ExpenseRow";
 
 const ExpenseTable = ({ expense }: { expense: Expense[] }) => {
-  const [filter, setFilter] = useState<keyof expense>("entryDate");
+  const [filter, setFilter] = useState<keyof Expense>("entryDate");
 
   const sortedExpenses = useMemo(() => {
     return sortExpenses(expense, filter);
@@ -21,7 +20,7 @@ const ExpenseTable = ({ expense }: { expense: Expense[] }) => {
         >
           <path d="M3.9 54.9C10.5 40.9 24.5 32 40 32H472c15.5 0 29.5 8.9 36.1 22.9s4.6 30.5-5.2 42.5L320 320.9V448c0 12.1-6.8 23.2-17.7 28.6s-23.8 4.3-33.5-3l-64-48c-8.1-6-12.8-15.5-12.8-25.6V320.9L9 97.3C-.7 85.4-2.8 68.8 3.9 54.9z" />
         </svg>
-        <select onChange={(e) => setFilter(e.target.value as keyof expense)}>
+        <select onChange={(e) => setFilter(e.target.value as keyof Expense)}>
           <option value="entryDate" label="Date" />
           <option value="amount" label="Amount" />
           <option value="category" label="Category" />
@@ -31,20 +30,7 @@ const ExpenseTable = ({ expense }: { expense: Expense[] }) => {
         <tbody>
           {sortedExpenses &&
             sortedExpenses.map((exp, index) => (
-              <tr key={index}>
-                <td className="expenseAmount">
-                  <RecurringIcon />$ {convertToFloat(exp.amount)}
-                </td>
-                <td>{exp.description}</td>
-                <td>{exp.category}</td>
-                <td>
-                  {new Date(exp.entryDate).toLocaleString("en-US", {
-                    month: "2-digit",
-                    day: "2-digit",
-                    year: "numeric",
-                  })}
-                </td>
-              </tr>
+              <ExpenseRow key={index} expense={exp} />
             ))}
         </tbody>
       </table>
