+++ src/app/dashboard/page.css
@@ -417,10 +417,6 @@ body {
   font-size: 0.9em;
 }
 
-.recurringIcon {
-  display: none;
-}
-
 /* Section colors */
 
 #Income.section
